requirements.txt
python-telegram-bot==21.6
openai>=1.0.0,<2.0.0
markdown-it-py>=3.0.0
linkify-it-py>=2.0.0
